Made get_link() and handle_parsed() simpler.
get_link() and handle_parsed() have been simplified.  Also, get_link()
will now download audio files and video files (for example, .mp3 files)
based on the file extension as well as the MIME type. This is useful
when the url doesn't have a mimetype, but it is still an MP3 file.